requirements.txt
amqp==2.5.2
billiard==3.6.1.0
celery==4.3.0
certifi==2019.11.28
chardet==3.0.4
coverage==4.5.4
Django==2.2.7
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-rest-framework==0.1.0
django-timezone-field==3.1
djangorestframework==3.10.3
idna==2.8
importlib-metadata==1.0.0
kombu==4.6.6
more-itertools==8.0.0
python-crontab==2.4.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
vine==1.3.0
zipp==0.6.0
